Eliminado SDLGameObject
Todo lo que contenía e implementaba está ahora en el 'GameObject'
padre, que ya no es abstracto y posee un .cpp

+++ src/Kepri2D/ObjectLayer.cpp
@@ -14,7 +14,7 @@ ObjectLayer::~ObjectLayer()
 void ObjectLayer::Update()
 {
 	// Actualizar los GameObjects
-	for (int i = 0; i < m_gameObjects.size(); i++)
+	for (size_t i = 0; i < m_gameObjects.size(); i++)
 	{
 		m_gameObjects[i]->Update();
 	}
@@ -22,7 +22,7 @@ void ObjectLayer::Update()
 void ObjectLayer::Render()
 {
 	// Pintar los GameObjects
-	for (int i = 0; i < m_gameObjects.size(); i++)
+	for (size_t i = 0; i < m_gameObjects.size(); i++)
 	{
 		m_gameObjects[i]->Draw();
 	}
